requirements.txt
# Minimal dependencies for LINE RAG Bot
pinecone[grpc]>=7.3.0
sentence-transformers>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
flask>=2.3.0
line-bot-sdk>=3.5.0